+++ packages/auto-complete/src/async-autoComplete.example.tsx
@@ -1,20 +1,35 @@
 import * as React from 'react';
-import AutoComplete from './AutoComplete';
+import AutoComplete, { DataSourceItem } from './AutoComplete';
 import { Phone } from '@lendi-ui/icon';
-import { MOCK_DATA_SOURCE } from './AutoComplete/MOCK_DATA_SOURCE';
+import { MOCK_DATA_SOURCE } from './AutoComplete/data-source.mock';
+import { pl } from '@lendi-ui/spacing';
+import styled from 'styled-components';
+
+const PaddedPhoneIcon = styled(Phone)`
+  ${pl('sm')}
+`;
 
 export default () => (
   <>
-    <strong>Async Input Drop Down(With Promise resolve)</strong>
-    <AutoComplete dataSource={(userInput: string) => getData(userInput)} />
+    <strong>Async Input Drop Down(With Promise resolve) and Fullwidth</strong>
+    <AutoComplete dataSource={(userInput: string) => getData(userInput)} isFullWidth={true} />
     <br />
     <br />
-    <hr />
+    <strong>With container max-width of 300px</strong>
+    <div style={{ maxWidth: '300px' }}>
+      <AutoComplete dataSource={(userInput: string) => getData(userInput)} />
+    </div>
+    <br />
+    <br />
     <strong>Before Icon</strong>
-    <AutoComplete dataSource={(userInput: string) => getData(userInput)} beforeIcon={<Phone color="shade.500" />} />
+    <AutoComplete
+      dataSource={(userInput: string) => getData(userInput)}
+      before={<PaddedPhoneIcon color="shade.500" />}
+      isFullWidth={true}
+    />
     <br />
     <br />
-    <strong>Example with all props like onSelect, size(lg) and placeholder</strong>
+    <strong>Example with some props like onSelect, size(lg) and placeholder</strong>
     <AutoComplete
       dataSource={(userInput: string) => getData(userInput)}
       onSelect={(text: string) => alert(text)}
@@ -33,16 +48,15 @@ export default () => (
     <br />
     <strong>Size 'sm'</strong>
     <AutoComplete size="sm" dataSource={(userInput: string) => getData(userInput)} placeholder={'Default md'} />
-
-    <br />
-    <br />
   </>
 );
 
-function getData(userInput: string): Promise<string[]> {
+function getData(userInput: string): Promise<DataSourceItem[]> {
   return new Promise((res, rej) => {
     setTimeout(() => {
-      const filteredDataSource = MOCK_DATA_SOURCE.filter((d) => d.toLowerCase().indexOf(userInput.toLowerCase()) > -1);
+      const filteredDataSource = MOCK_DATA_SOURCE.filter(
+        (data) => data.label.toLowerCase().indexOf(userInput.toLowerCase()) > -1
+      );
       return res(filteredDataSource);
     }, 400);
   });