+++ packages/auto-complete/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [3.0.0] - 2019-05-20 - [HUB-172](https://creditandfinance.atlassian.net/browse/HUB-172)
+### Changed
+- change the datasource prop, now it can take an Object<label, value>
+- Full width support
+- can take value as props
+- added few more props like isError, className, isInverse, isRequired, isReadOnly, isAutoFocus
+
 ## [2.1.0] - 2019-05-17 - [HUB-174](https://creditandfinance.atlassian.net/browse/HUB-174)
 ### Changed
 - Module format reverted back to CJS.