+++ packages/text-input/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [3.1.1] - 2019-05-20 - [HUB-172](https://creditandfinance.atlassian.net/browse/HUB-172)
+### Changed
+- Fix a icon vertical-aligning
+
 ## [3.1.0] - 2019-05-17 - [HUB-174](https://creditandfinance.atlassian.net/browse/HUB-174)
 ### Changed
 - Module format reverted back to CJS.